+++ addon/components/paper-menu-container.js
@@ -1,0 +1,6 @@
+import PaperMenuContainerAbstract from './paper-menu-container-abstract';
+
+export default PaperMenuContainerAbstract.extend({
+  classNames: ['md-whiteframe-z2','md-open-menu-container'],
+  interaction: true
+});