+++ addon/components/paper-menu-item.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  tagName: 'md-menu-item',
+
+  actions: {
+    action() {
+      this.sendAction('action', this.get('param'));
+    }
+  }
+
+});