+++ addon/components/paper-select-container.js
@@ -1,0 +1,6 @@
+import PaperMenuContainerAbstract from './paper-menu-container-abstract';
+
+export default PaperMenuContainerAbstract.extend({
+  classNames: ['md-select-menu-container'],
+  interaction: true
+});